requirements.txt
openai>=1.30.0
requests>=2.31.0
beautifulsoup4>=4.12.0
playwright>=1.40.0
python-dotenv>=1.0.0
